config: only delete records if there is more than one for a given field name, re #6222
git-svn-id: svn://develop.studip.de/studip/trunk@36388 c5b85792-5220-0410-ad23-ed5b6b8ad999

+++ lib/classes/Config.class.php
@@ -301,7 +301,7 @@ class Config implements ArrayAccess, Countable, IteratorAggregate
             // store the default-type for the modified entry
             $entry->type = $this->metadata[$field]['type'];
 
-            if (!$entry->is_default && $entry->value == $entries[0]->value) {
+            if (count($entries) > 1 && !$entry->is_default && $entry->value == $entries[0]->value) {
                 $ret += $entry->delete();
             } else {
                 $ret += $entry->store();
